Refactor transport into multiple modules.
Addresses #158.

Besides the refactoring, the type `EitherIncomingStream` has been
removed as it is not used.
