+++ swarm/src/upgrade/plaintext.rs
@@ -1,0 +1,54 @@
+// Copyright 2017 Parity Technologies (UK) Ltd.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a
+// copy of this software and associated documentation files (the "Software"),
+// to deal in the Software without restriction, including without limitation
+// the rights to use, copy, modify, merge, publish, distribute, sublicense,
+// and/or sell copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+// DEALINGS IN THE SOFTWARE.
+
+use bytes::Bytes;
+use futures::future::{self, FutureResult};
+use multiaddr::Multiaddr;
+use std::{iter, io::Error as IoError};
+use tokio_io::{AsyncRead, AsyncWrite};
+use upgrade::{ConnectionUpgrade, Endpoint};
+
+/// Implementation of the `ConnectionUpgrade` that negotiates the `/plaintext/1.0.0` protocol and
+/// simply passes communications through without doing anything more.
+///
+/// > **Note**: Generally used as an alternative to `secio` if a security layer is not desirable.
+// TODO: move to a separate crate?
+#[derive(Debug, Copy, Clone)]
+pub struct PlainTextConfig;
+
+impl<C> ConnectionUpgrade<C> for PlainTextConfig
+where
+    C: AsyncRead + AsyncWrite,
+{
+    type Output = C;
+    type Future = FutureResult<C, IoError>;
+    type UpgradeIdentifier = ();
+    type NamesIter = iter::Once<(Bytes, ())>;
+
+    #[inline]
+    fn upgrade(self, i: C, _: (), _: Endpoint, _: &Multiaddr) -> Self::Future {
+        future::ok(i)
+    }
+
+    #[inline]
+    fn protocol_names(&self) -> Self::NamesIter {
+        iter::once((Bytes::from("/plaintext/1.0.0"), ()))
+    }
+}
